hotFix: "PointServiceImpl 삭제"
- Interface 안써서 과감히 삭제
